- update coroutines to 1.0.1 - PulseLed.durationMillis from int to long

+++ app/src/main/java/com/pierdr/tramontana/io/websocket/Protocol.kt
@@ -26,7 +26,7 @@ class Protocol {
             "transitionColors" -> parseTransitionColors(json)
             "setBrightness" -> Directive.SetBrightness(json.getFloat("b"))
             "setLED" -> Directive.SetLed(json.getFloat("in"))
-            "pulseLED" -> Directive.PulseLed(json.getInt("t"), (json.getFloat("d") * 1000).toInt(), 1f)
+            "pulseLED" -> Directive.PulseLed(json.getInt("t"), (json.getFloat("d") * 1000).toLong(), 1f)
             "showImage" -> Directive.ShowImage(json.getString("url"))
             "playVideo" -> Directive.PlayVideo(json.getString("url"))
             "playAudio" -> Directive.PlayAudio(json.getString("url"))
